requirements.txt
# Data generation
faker==24.0.0
kafka-python==2.0.2

# Spark + Delta Lake
pyspark==3.5.1
delta-spark==3.1.0

# AWS
boto3==1.34.0
s3fs==2024.2.0

# GCP / BigQuery
google-cloud-bigquery==3.17.0
google-cloud-storage==2.14.0
db-dtypes==1.2.0

# Airflow (install separately — see guide)
# apache-airflow==2.8.0
# apache-airflow-providers-amazon==8.18.0
# apache-airflow-providers-google==10.14.0

# dbt
dbt-core==1.7.0
dbt-bigquery==1.7.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
